// src/Utilities/Allstate/models/EnrollmentApplicationQuestion.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Group Markets API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import {
    EnrollmentResponseType,
} from './';

/**
 * 
 * @export
 * @interface EnrollmentApplicationQuestion
 */
export interface EnrollmentApplicationQuestion {
    /**
     * Defines the property name to be used in the submit application endpoint payload
     * @type {string}
     * @memberof EnrollmentApplicationQuestion
     */
    Code?: string | null;
    /**
     * 
     * @type {string}
     * @memberof EnrollmentApplicationQuestion
     */
    Description?: string | null;
    /**
     * 
     * @type {EnrollmentResponseType}
     * @memberof EnrollmentApplicationQuestion
     */
    ResponseType?: EnrollmentResponseType;
}