// src/Utilities/Allstate/models/EnrollmentEmployeeStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Group Markets API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum EnrollmentEmployeeStatus {
    W2 = 'W2',
    F1099 = 'F1099',
    OwnerPartner = 'OwnerPartner'
}